pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pooxm-universe-map"
version = "0.1.0"
description = "CUDA PyTorch physics demo: toy self-gravitating system producing a 'map of the universe'."
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "Your Name" }]
dependencies = [
    "torch>=2.2",
    "numpy>=1.24",
    "matplotlib>=3.7",
    "PyYAML>=6.0",
    "tqdm>=4.66",
]

[project.urls]
Homepage = "https://example.com"

[project.scripts]
pooxm = "pooxm.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["pooxm*"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q"
pythonpath = ["src"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP"]
ignore = []
